backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-multipart==0.0.18
pyyaml==6.0.2
pytest==8.3.4
